requirements.txt
langchain
langchain-anthropic
langchain-community
wikipedia
python-dotenv
pydantic
duckduckgo-search


